aisp-poc/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
pydantic>=2.7
supabase-py>=2.4
openai>=1.25
python-dotenv>=1.0
jinja2>=3.1
sqlmodel>=0.0.16
jsonschema>=4.22
httpx>=0.27